De-dupe key check logic into common function
